// packages/helper/src/components/card/card-types.ts
export type CardVariant =
  | 'elevated'
  | 'outlined'
  | 'flat'
  | 'tonal'
  | 'glass'

export type CardColor =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'teal'
  | 'purple'
  | 'dark'
  | 'neutral'
  | 'light'